+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ==================
 
 - use FreeBSD 10.0 as default for bootstrapping and documentation
+- always encode result of templates as utf-8
 
 
 1.0.0 - 2014-07-20